refactor(telegram): extract callback-query handlers from gateway.ts (#2996)
Phase 5 remaining item 2: move the vault / skill / mental-model /
operator-event / auth-dashboard callback handler families (~2,200 lines)
verbatim out of gateway.ts into callback-query-handlers.ts behind a
createCallbackQueryHandlers(deps) factory, following the extraction style
of PRs #3007-#3011. Behavior-preserving: bodies are byte-identical except
four mechanical reads of mutable module lets that became injected getters.
